+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 
 ## Neste versjon
 
+- ✨ **Changelog**. Expand bokser i changelog for å se tidligere endringer/versjoner.
+
 ## Versjon 1.2.3 (29.09.2021)
 
 - ✨ **Varslinger**. Varsler har blitt flyttet til topbaren slik at man lettere ser nye varsler.